test: toJSON works properly on booleans
